% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeLandcover_to_ChangeDepVar.R
\name{ChangeLandcover_to_ChangeDepVar}
\alias{ChangeLandcover_to_ChangeDepVar}
\title{Run simulation of landcover spread and its effect on a dependent variable}
\usage{
ChangeLandcover_to_ChangeDepVar(
  landcover_list,
  infest_val,
  suscep_val,
  dep_var_raster_initial,
  dep_var_raster_pred,
  dep_var_modifier
)
}
\arguments{
\item{landcover_list}{A list of yearly landcover rasters, typically from the LandCoverSpread function}

\item{infest_val}{numerical. Value of the landcover that will be spreading.}

\item{suscep_val}{numerical. Vector of landcover values that are susceptible to the spread.}

\item{dep_var_raster_initial}{raster. The initial raster for the dependent variable. It is recommended you use a raster of predicted values rather than raw data (i.e. use \code{gls_spatial_predict()} with \code{landcover_val == 'ALL'} specified).}

\item{dep_var_raster_pred}{raster. A raster with predicted values for the dependent variable in all pixels susceptible to invasion. Usually a raster from \code{gls_spatial_predict()}.}

\item{dep_var_modifier}{numerical. A scalar to optionally return a list of rasters with modified dep_var rasters (e.g. multiply water yield rasters to obtain recharge rasters)}
}
\value{
TO DO
}
\description{
This function runs a simulation of landcover spread, using bootstrapped results. It returns a list of landcover rasters and a list of the rasters for your dependent variable; one for each year.
}
\details{
TO DO
}
\examples{
TO DO

}
