% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSubset.R
\name{datSubset}
\alias{datSubset}
\title{Run spatial GLS models with optimal spatial correlation structure.}
\usage{
datSubset(
  data,
  x,
  y,
  shp_reg,
  shp_app = NULL,
  sample = NULL,
  convertFromUTM = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} that can be converted to a spatial \code{data.frame}. Coordinates must be in latitude and longitude (decimal degrees)}

\item{x}{chr string specifying the x-coordinates in \code{data}}

\item{y}{chr string specifying the y-coordinates in \code{data}}

\item{shp_reg}{\code{shapefile} outlining the area of \code{data} to be used for the regression}

\item{shp_app}{\code{shapefile} outlining the area of \code{data} to which the final simulation will be applied, if different from \code{shp_reg}}

\item{sample}{integer. If the regression data are large, you may wish to sample the rows for the \code{gls_spatial} function.}

\item{convertFromUTM}{logical. Set to \code{TRUE} if your shapfile(s) are in UTM coordinates.}
}
\value{
A list of two \code{data.frames}. The first is a \code{data.frame} with all points subsetted using the \code{shp_reg} shapefile. The second is a \code{data.frame} with all points subsetted using the \code{shp_app} or \code{shp_reg} shapefile, as appropriate.
}
\description{
This function runs a GLS model on a spatial dataset. It chooses among five types of correlation structure, and returns the model with the lowest AIC.
}
\details{
Subsets a \code{data.frame} according to provided \code{shapefiles}. The
}
\examples{

}
