% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSimulation.R
\name{fullSimulation}
\alias{fullSimulation}
\title{Run full simulation, accepting all defaults}
\usage{
fullSimulation(
  data,
  shp_reg = NULL,
  shp_app = NULL,
  convertFromUTM = FALSE,
  dat_sample = NULL,
  landcover_varname,
  reg_formula,
  landcover_invasive,
  landcover_susceptible,
  dep_varname,
  x_coords_varname,
  y_coords_varname,
  spread_rate,
  birdcell,
  simlength,
  simulation_count = 1000,
  dep_var_modifier,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{\code{data.frame} with spatial data}

\item{shp_reg}{\code{shapefile} outlining the area of \code{data} to be used for the regression}

\item{shp_app}{\code{shapefile} outlining the area of \code{data} to which the final simulation will be applied, if different from \code{shp_reg}}

\item{convertFromUTM}{logical. Set to \code{TRUE} if you are subsetting your data with shapefile(s) and your shapefile(s) are in UTM coordinates instead of lon/lat.}

\item{dat_sample}{numerical. If specified, will take a sample of \code{n} observations for the regression (suggested if \code{shp_reg} has lots of observations)}

\item{landcover_varname}{character string specifying the landcover variable from \code{data}}

\item{reg_formula}{regression formula to be used, as in \code{lm()}, with columns from \code{data} (e.g. \code{c ~ a + b})}

\item{landcover_invasive}{value. Numerical or character value of the invasive landcover.}

\item{landcover_susceptible}{value. Numerical or character value(s) of the susceptible landcover(s). If more than one susceptible landcover, provide a vector \code{c()}.}

\item{x_coords_varname}{character string. Name of the x-coordinate variable in \code{data}.}

\item{y_coords_varname}{character string. Name of the y-coordinate variable in \code{data}.}

\item{spread_rate}{numerical. Value between 0 and 1 indicating the annual spread rate of the invading landcover.}

\item{birdcell}{numerical. Value between 0 and 1 indicating the probability a random cell can be invaded, regardless of adjacency to existing invaded pixels.}

\item{simlength}{integer. Number of years the simulation should run.}

\item{simulation_count}{integer. Length of simulation bootstrap.}

\item{dep_var_modifier}{numerical. A scalar to optionally return a list of rasters with modified dep_var rasters (e.g. multiply water yield rasters to obtain recharge rasters)}

\item{num_cores}{numerical. Number of cores for parallel processing, max 5}
}
\value{
A list of all objects returned by the individual functions
}
\description{
This function runs the entire simulation by wrapping all other functions into a single function, accepting all default customization options
}
\details{

}
\examples{
TO DO

}
