% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_converter.R
\name{fullSimulation_unit_converter}
\alias{fullSimulation_unit_converter}
\title{Convert the units of a LandCoverSpread simulation}
\usage{
fullSimulation_unit_converter(
  fullSimulation_results,
  unit_conversion_factor,
  dep_var_modified = TRUE
)
}
\arguments{
\item{fullSimulation_results}{list. Object returned from the \code{fullSimulation} or \code{fullSimulationApp} function.}

\item{unit_conversion_factor}{numerical. The value to be multiplied with all rasters for unit conversion.}

\item{dep_var_modified}{logical. Indicates whether the \code{dep_var} was modified in \code{LandCoverSpread}}
}
\value{
A list with the same length and contents as \code{fullSimulation}.
}
\description{
This function converts the units of all rasters (except landcover rasters) returned from the \code{LandCoverSpread} function.
}
\details{
This function simply multiplies the values of the values of the results of the simulation. If you modified the dependent variable in the simulation, you should specify \code{dep_var_modified = TRUE}.
}
\examples{

}
