% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSimulationApp.R
\name{fullSimulationApp}
\alias{fullSimulationApp}
\title{Run full simulation, accepting all defaults}
\usage{
fullSimulationApp(
  data,
  shp_reg = NULL,
  shp_app = NULL,
  convertFromUTM = FALSE,
  dat_sample = NULL,
  landcover_varname,
  reg_formula,
  landcover_invasive,
  landcover_susceptible,
  x_coords_varname,
  y_coords_varname,
  spread_rate,
  birdcell,
  simlength,
  simulation_count = 100,
  dep_var_modifier = NA,
  unit_converter = 1,
  covar_adjustment = 1,
  zero_break = FALSE,
  outlier_value = NA,
  dep_var_plot_label = "mm/yr",
  dep_var_plot_label_cumulative = "mm",
  line_plot_labels = c("Water yield", "Recharge"),
  line_plot_positive_vals_only = TRUE,
  line_plot_axis_label = "mm/yr",
  landcover_label_invasive = "Invasive",
  landcover_label_susceptible = "Susceptible",
  preview_plot_scalebar_position = "bottomright",
  preview_plot_scalebar_unit = "km",
  preview_plot_scalebar_dist = 10,
  priority_plot_scalebar_position = "bottomright",
  priority_plot_scalebar_unit = "km",
  priority_plot_scalebar_dist = 1,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{\code{data.frame} if data is in environement, or character string specifying location of xlsx data with directory if \code{data_as_directories = TRUE}}

\item{shp_reg}{character string specifying layer (.shp) filename corresponding to \code{shp_reg_directory}, or spatial polygon object if \code{data_as_directories = FALSE}}

\item{shp_app}{character string specifying layer (.shp) filename corresponding to \code{shp_app_directory}, or spatial polygon object if \code{data_as_directories = FALSE}}

\item{convertFromUTM}{logical. Set to \code{TRUE} if you are subsetting your data with shapefile(s) and your shapefile(s) are in UTM coordinates instead of lon/lat.}

\item{dat_sample}{numerical. If specified, will take a sample of \code{n} observations for the regression (suggested if \code{shp_reg} has lots of observations)}

\item{landcover_varname}{character string specifying the landcover variable from \code{data}}

\item{reg_formula}{regression formula to be used, as in \code{lm()}, with columns from \code{data} (e.g. \code{c ~ a + b})}

\item{landcover_invasive}{value. Numerical or character value of the invasive landcover.}

\item{landcover_susceptible}{value. Numerical or character value(s) of the susceptible landcover(s). If more than one susceptible landcover, provide a vector \code{c()}.}

\item{x_coords_varname}{character string. Name of the x-coordinate variable in \code{data}.}

\item{y_coords_varname}{character string. Name of the y-coordinate variable in \code{data}.}

\item{spread_rate}{numerical. Value between 0 and 1 indicating the annual spread rate of the invading landcover.}

\item{birdcell}{numerical. Value between 0 and 1 indicating the probability a random cell can be invaded, regardless of adjacency to existing invaded pixels.}

\item{simlength}{integer. Number of years the simulation should run.}

\item{simulation_count}{integer. Length of simulation bootstrap.}

\item{dep_var_modifier}{numerical. A scalar to optionally return a list of rasters with modified dep_var rasters (e.g. multiply water yield rasters to obtain recharge rasters)}

\item{unit_converter}{numerical. A scalar to optionally modify the values of the resulting dependent variable (and modified dependent variable, if present) to convert units (e.g. mm/yr to gal/acre/day). Default value is \code{1}.}

\item{covar_adjustment}{list. List specifying change in covariate values. See ?gls_spatial_predict.}

\item{zero_break}{numerical. Should the priority plots have a breakpoint at 0 if values are both negative and positive?}

\item{outlier_value}{numerical. Cutoff value for outliers in the priority plots.}

\item{dep_var_plot_label}{character. Label for the dependent variable used in plots.}

\item{dep_var_plot_label_cumulative}{character. Legend title for the cumulative change priority plot (also applied to modified values, if \code{dep_var_modifier} provided).}

\item{line_plot_labels}{character vector of length 1 or 2, for dependent variable and modified dependent variable if provided, for the line graph. e.g. \code{c('Water yield', 'Recharge')}}

\item{line_plot_positive_vals_only}{logical. When plotting the line graph, should the negative values be removed when aggregating by year?}

\item{num_cores}{numerical. Number of cores for parallel processing, max 5}

\item{data_as_directories}{logical. \code{TRUE} if data and shapefiles must be loaded from files. \code{FALSE} if data and shapefiles are \code{R} objects already loaded in the environment.}

\item{shp_reg_directory}{character string specifying directory of shapefile outlining the area of \code{data} to be used for the regression. \code{NULL} if \code{data_as_directories = FALSE}}

\item{shp_app_directory}{character string specifying directory of shapefile outlining the area of \code{data} to which the final simulation will be applied, if different from \code{shp_reg}.  \code{NULL} if \code{data_as_directories = FALSE}}
}
\value{
A list of all objects returned by the individual functions
}
\description{
This function runs the entire simulation by wrapping all other functions into a single function, accepting all default customization options
}
\details{

}
\examples{
TO DO

}
