% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSimulation.R
\name{fullSimulation}
\alias{fullSimulation}
\title{Run full simulation, accepting all defaults}
\usage{
fullSimulation(
  data,
  landcover_varname,
  landcover_vec,
  reg_formula,
  error_formula,
  landcover_invasive,
  landcover_susceptible,
  dep_varname,
  x_coords_varname,
  y_coords_varname,
  spread_rate,
  birdcell,
  simlength,
  simulation_count,
  dep_var_modifier,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{\code{data.frame} with spatial data}

\item{landcover_varname}{character string specifying the landcover variable from \code{data}}

\item{landcover_vec}{vector of all landcover types in \code{data$landcover_varname} to be analyzed. This should include the landcover that is spreading along with all landcovers that are susceptible to the spread.}

\item{reg_formula}{regression formula to be used, as in \code{lm()}, with columns from \code{data} (e.g. \code{c ~ a + b})}

\item{error_formula}{one-sided formula specifying coordinate columns from \code{data} (e.g. \code{~ x + y})}

\item{landcover_invasive}{value. Numerical or character value of the invasive landcover.}

\item{landcover_susceptible}{value. Numerical or character value(s) of the susceptible landcover(s). If more than one susceptible landcover, provide a vector \code{c()}.}

\item{dep_varname}{character string. Name of the dependent variable in \code{data}.}

\item{x_coords_varname}{character string. Name of the x-coordinate variable in \code{data}.}

\item{y_coords_varname}{character string. Name of the y-coordinate variable in \code{data}.}

\item{spread_rate}{numerical. Value between 0 and 1 indicating the annual spread rate of the invading landcover.}

\item{birdcell}{numerical. Value between 0 and 1 indicating the probability a random cell can be invaded, regardless of adjacency to existing invaded pixels.}

\item{simlength}{integer. Number of years the simulation should run.}

\item{simulation_count}{integer. Length of simulation bootstrap.}

\item{dep_var_modifier}{numerical. A scalar to optionally return a list of rasters with modified dep_var rasters (e.g. multiply water yield rasters to obtain recharge rasters)}

\item{num_cores}{numerical. Number of cores for parallel processing, max 5}
}
\value{
A list of all objects returned by the individual functions
}
\description{
This function runs the entire simulation by wrapping all other functions into a single function, accepting all default customization options
}
\details{

}
\examples{
TO DO

}
