% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverSpread.R
\name{LandCoverSpread}
\alias{LandCoverSpread}
\title{Run simulation of landcover spread and its effect on a dependent variable}
\usage{
LandCoverSpread(
  infest_val,
  suscep_val,
  spread_rate,
  birdcell,
  simlength,
  simulation_count,
  lc_raster,
  dep_var_raster_initial,
  dep_var_raster_pred,
  dep_var_modifier = NA,
  silent = FALSE
)
}
\arguments{
\item{infest_val}{numerical. Value of the landcover that will be spreading.}

\item{suscep_val}{numerical. Vector of landcover values that are susceptible to the spread.}

\item{spread_rate}{numerical. Value between 0 and 1 indicating the annual spread rate of the invading landcover.}

\item{birdcell}{numerical. Value between 0 and 1 indicating the probability a random cell can be invaded, regardless of adjacency to existing invaded pixels.}

\item{simlength}{integer. Number of years the simulation should run.}

\item{simulation_count}{integer. Length of simulation bootstrap.}

\item{lc_raster}{raster. The initial landcover raster.}

\item{dep_var_raster_initial}{raster. The initial raster for the dependent variable. It is recommended you use a raster of predicted values rather than raw data (i.e. use \code{gls_spatial_predict()} with \code{landcover_val == 'ALL'} specified).}

\item{dep_var_raster_pred}{raster. A raster with predicted values for the dependent variable in all pixels susceptible to invasion. Usually a raster from \code{gls_spatial_predict()}.}

\item{dep_var_modifier}{numerical. A scalar to optionally return a list of rasters with modified dep_var rasters (e.g. multiply water yield rasters to obtain recharge rasters)}

\item{silent}{logical. Suppress printing notification when each simulation is done. Default is \code{FALSE}.}
}
\value{
Lists of rasters (landcover and dependent variable) and, optionally, raster files to a directory of your choosing.
}
\description{
This function runs a simulation of landcover spread, using bootstrapped results. It returns a list of landcover rasters and a list of the rasters for your dependent variable; one for each year.
}
\details{
The output is a list with two elements: the first element is a list of landcover rasters from year 0 (the initial raster) to year n (the length of the spread period). The second element is a list of rasters for your dependent variable.
}
\examples{
# load packages
library(LandCover); library(foreach)

# initialize data.frame with coordinates
dat <- expand.grid(x = 1:20, y = 1:20, KEEP.OUT.ATTRS = FALSE)

# create some data: elevation, landcover, and temp/ET dependent on elevation and landcover
dat$elevation <- with(dat, 50 + 2*x + 5*y + rnorm(nrow(dat), sd = 7))
dat$landcover <- ifelse(dat$elevation < median(dat$elevation), 1, 2)
dat$temp      <- with(dat, (120-0.7*(0.5*elevation + 0.3*y - 0.5*x + ifelse(landcover == 'lc1', -30, 0) + rnorm(nrow(dat)))))
dat$ET        <- with(dat, (   -0.4*(-2*temp       + 0.5*y - 1.0*x + ifelse(landcover == 'lc1', +20, 0) + rnorm(nrow(dat)))))

# run the gls model
regression_results <- gls_spatial(data = dat, landcover_varname = 'landcover', landcover_vec = c(1,2),
                                  reg_formula = ET ~ elevation + temp, error_formula = ~ x + y)

}
